+++ src/Api/Exception/ClientException.php
@@ -1,0 +1,8 @@
+<?php
+declare(strict_types=1);
+
+namespace Mxm\Api\Exception;
+
+class ClientException extends RuntimeException
+{
+}